requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15.2
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.3
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
filelock==3.7.1
google-api-python-client==1.6.4
httplib2==0.20.4
idna==3.3
jsonpickle==0.9.6
oauth2client==4.1.2
oauthlib==3.2.1
phonenumberslite==8.12.54
Pillow==9.2.0
pipenv==2022.6.7
platformdirs==2.5.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
uritemplate==3.0.1
urllib3==1.26.12
virtualenv==20.14.1
virtualenv-clone==0.5.7
